pyproject.toml
[project]
name = "tg-ticket-provider"
version = "0.1.0"
description = "Telegram bot for internal ticket workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13,<4",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["tg_ticket_provider*"]
